libs: Changed the scaling library to make use of the games calculated window and menu resolutions instead of using the values from the current boot configurations.
This should fix issues with users setting the window or menu resolutions to invalid values or using values such as 0 or -1 to allow the game to make use of its own defaults. This will now cause the scaling library to pull the actual in-use values from the internal CYmdb manager.