New Feature: Additional Keybind Modifiers
This changeset introduces a new set of key modifiers that can be used with key binds.

 - '$' - The key bind only works when game input is active / opened.
 - '%' - The key bind only works when game input is inactive / closed.

Previously, key binds would always fire regardless of the game input state, unless the user has adjusted the configuration value 'keyboard.blockbindsduringinput'. These new modifiers allow key binds to be more explicit with their usage now.

Additionally, the configuration value 'keyboard.blockbindsduringinput' has been changed to default off. This configuration value is now used as a global override to all keybinds to be blocked when enabled and input is open/active, regardless if they make use of these new modifiers. (It is now recommended that players disable this configuration option and be more explicit with their binds.)

 - sdk: Updated project copyright year.
 - sdk: Updated IKeyboard function prototypes for keybinds.
 - proj: Updated default script to adjust the /paste keybind.